LayoutTests/contact-picker/contacts-manager-get-properties.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContactPickerAPIEnabled=true ] -->
<html>
    <head>
        <script src="../resources/js-test.js"></script>
    </head>
    <script>
        jsTestIsAsync = true;

        async function runTest()
        {
            description("Test for ContactsManager.getProperties().");

            properties = await navigator.contacts.getProperties();
            shouldBeType("properties", "Array");
            shouldBeEqualToNumber("properties.length", 3);
            shouldBeTrue("properties.includes('email')");
            shouldBeTrue("properties.includes('name')");
            shouldBeTrue("properties.includes('tel')");

            finishJSTest();
        }
    </script>
    <body onload=runTest()></body>
</html>
