Safari-611.1.10/LayoutTests/contact-picker/contacts-manager-get-properties-expected.txt

Test for ContactsManager.getProperties().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS properties is an instance of Array
PASS properties.length is 3
PASS properties.includes('email') is true
PASS properties.includes('name') is true
PASS properties.includes('tel') is true
PASS successfullyParsed is true

TEST COMPLETE
